Effectively split up simulator and simulation context.
